requirements.txt
django
django-cors-headers
django-cockroachdb
djangorestframework
psycopg2
psycopg2-binary